+++ src/models/action_net_2.py
@@ -4,8 +4,8 @@ from typing import Any
 
 import pytorch_lightning as pl
 import torch
-import torch.nn.functional as F
 import yaml
+from Convo_trainer import ActionTraining
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.loggers import TensorBoardLogger
 from torch import nn
@@ -17,10 +17,11 @@ if str(PROJECT_ROOT) not in sys.path:
 
 from pathlib import Path
 
+from Convo_model import ActionNet
+
 from src.dataset.pre_process import ComposePreprocessor, Resize, StackWithLabels
 from src.dataset.torch_dataset import get_torch_dataloaders
 from src.Eval.gym_eval import GymManager, RecordingGymManager, RuntimePreprocessor
-from src.models.utils import atari_to_gym
 from src.utils import skip_run
 
 # from src.utils skip_run
@@ -113,256 +114,6 @@ class ActionNet(nn.Module):
         )
 
         return x
-
-
-# ================================================================
-# Lightning training module
-# ================================================================
-
-
-class ActionTraining(pl.LightningModule):
-    def __init__(
-        self,
-        hparams: dict,
-        net: nn.Module,
-        data_loader: dict,
-    ):
-        super().__init__()
-
-        self.net = net
-        self.data_loaders = data_loader
-
-        self.learning_rate = float(hparams.get("learning_rate", 1e-3))
-        self.weight_decay = float(hparams.get("weight_decay", 0.0))
-
-        self.save_hyperparameters(
-            {
-                "learning_rate": self.learning_rate,
-                "weight_decay": self.weight_decay,
-                "num_actions": hparams["num_actions"],
-            }
-        )
-
-    def forward(self, images: torch.Tensor) -> torch.Tensor:
-        return self.net(images)
-
-    @staticmethod
-    def _prepare_images(images: torch.Tensor) -> torch.Tensor:
-        """
-        Accept either:
-
-            [batch, 4, 84, 84]
-
-        or:
-
-            [batch, 4, 1, 84, 84]
-
-        ActionNet requires [batch, 4, 84, 84].
-        """
-        images = images.float()
-
-        if images.ndim == 5 and images.shape[2] == 1:
-            images = images.squeeze(2)
-
-        if images.ndim != 4:
-            raise ValueError(
-                "Training images must have shape "
-                "[batch, frames, height, width], but received "
-                f"{tuple(images.shape)}."
-            )
-
-        if images.shape[1] != 4:
-            raise ValueError(
-                "ActionNet requires four stacked frames, but received "
-                f"shape {tuple(images.shape)}."
-            )
-
-        return images
-
-    @staticmethod
-    def _prepare_actions(actions: torch.Tensor) -> torch.Tensor:
-        """
-        Prepare four action targets per sample.
-
-        Input:
-            [batch, 4]
-            [batch, 4, 1]
-
-        Output:
-            [batch, 4] containing Gym action IDs 0-8.
-        """
-        actions = torch.as_tensor(actions)
-
-        if actions.ndim == 3 and actions.shape[-1] == 1:
-            actions = actions.squeeze(-1)
-
-        if actions.ndim != 2:
-            raise ValueError(
-                "Expected actions with shape [batch, 4], "
-                f"but received {tuple(actions.shape)}."
-            )
-
-        if actions.shape[1] != 4:
-            raise ValueError(
-                "Expected four actions per sample, "
-                f"but received shape {tuple(actions.shape)}."
-            )
-
-        actions = atari_to_gym(actions)
-
-        return actions.long()
-
-    def _shared_step(
-        self,
-        batch,
-        stage: str,
-    ) -> torch.Tensor:
-        images, _, actions = batch
-
-        images = self._prepare_images(images)
-        targets = self._prepare_actions(actions)
-
-        logits = self.net(images)
-
-        # Expected:
-        # logits:  [batch, 4, 9]
-        # targets: [batch, 4]
-        if logits.ndim != 3:
-            raise ValueError(
-                "Expected logits with shape [batch, 4, 9], "
-                f"but received {tuple(logits.shape)}."
-            )
-
-        if logits.shape[:2] != targets.shape:
-            raise ValueError(
-                "Prediction and target shapes do not match: "
-                f"logits={tuple(logits.shape)}, "
-                f"targets={tuple(targets.shape)}."
-            )
-
-        if logits.shape[2] != 9:
-            raise ValueError(
-                f"Expected nine Gym action classes, but received {logits.shape[2]}."
-            )
-
-        loss = F.cross_entropy(
-            logits.reshape(-1, 9),
-            targets.reshape(-1),
-        )
-
-        predicted_actions = logits.argmax(dim=2)
-
-        accuracy = (predicted_actions == targets).float().mean()
-
-        self.log(
-            f"{stage}_loss",
-            loss,
-            on_step=stage == "train",
-            on_epoch=True,
-            prog_bar=True,
-            logger=True,
-            batch_size=images.shape[0],
-        )
-
-        self.log(
-            f"{stage}_accuracy",
-            accuracy,
-            on_step=False,
-            on_epoch=True,
-            prog_bar=True,
-            logger=True,
-            batch_size=images.shape[0],
-        )
-
-        return loss
-
-    """
-    def _shared_step(
-        self,
-        batch,
-        stage: str,
-    ) -> torch.Tensor:
-        images, _, actions = batch
-
-        images = self._prepare_images(images)
-        targets = self._prepare_actions(actions)
-
-        logits = self.net(images)
-
-        if logits.shape[0] != targets.shape[0]:
-            raise ValueError(
-                "Batch size mismatch between predictions and targets: "
-                f"logits={tuple(logits.shape)}, "
-                f"targets={tuple(targets.shape)}."
-            )
-
-        if logits.shape[1] != 9:
-            raise ValueError(
-                "The network must produce nine Gym action logits, "
-                f"but produced shape {tuple(logits.shape)}."
-            )
-
-        loss = F.cross_entropy(logits, targets)
-
-        predicted_actions = logits.argmax(dim=1)
-        accuracy = (predicted_actions == targets).float().mean()
-
-        self.log(
-            f"{stage}_loss",
-            loss,
-            on_step=stage == "train",
-            on_epoch=True,
-            prog_bar=True,
-            logger=True,
-            batch_size=images.shape[0],
-        )
-
-        self.log(
-            f"{stage}_accuracy",
-            accuracy,
-            on_step=False,
-            on_epoch=True,
-            prog_bar=True,
-            logger=True,
-            batch_size=images.shape[0],
-        )
-
-        return loss"""
-
-    def training_step(self, batch, batch_idx):
-        return self._shared_step(batch, "train")
-
-    def validation_step(self, batch, batch_idx):
-        return self._shared_step(batch, "val")
-
-    def test_step(self, batch, batch_idx):
-        return self._shared_step(batch, "test")
-
-    def configure_optimizers(self):
-        return torch.optim.Adam(
-            self.net.parameters(),
-            lr=self.learning_rate,
-            weight_decay=self.weight_decay,
-        )
-
-    def train_dataloader(self):
-        return self.data_loaders["train"]
-
-    def val_dataloader(self):
-        loader = self.data_loaders.get("val")
-
-        if loader is None:
-            return []
-
-        return loader
-
-    def test_dataloader(self):
-        loader = self.data_loaders.get("test")
-
-        if loader is None:
-            return []
-
-        return loader
 
 
 # ================================================================
@@ -559,7 +310,7 @@ class CheckpointActionNet:
 
 
 # ================================================================
-# Skip run 1: train and save the network
+# Skip run 1: train the convo network and save the network
 # ================================================================
 
 with (
@@ -668,7 +419,7 @@ with (
 
 
 # ================================================================
-# Skip run 2: load the checkpoint and run one Gym episode
+# Skip run 2: load the checkpoint of the convo network and run one Gym episode
 # ================================================================
 for x in range(10):
     with (
@@ -755,7 +506,7 @@ with (
     max_steps = int(
         runtime_config.get(
             "gym_max_steps",
-            100_000,
+            100_00,
         )
     )
 
@@ -770,6 +521,7 @@ with (
         print("Initial state shape:", tuple(state.shape))
 
         done = False
+        previous_lives = None
 
         while not done and step_count < max_steps:
             state, reward, done, info = manager.step()
@@ -777,6 +529,16 @@ with (
             total_reward += float(reward)
             step_count += 1
             final_info = info
+
+            lives = info.get("lives")
+
+            if lives is not None:
+                if previous_lives is not None and lives < previous_lives:
+                    print(f"\nDeath detected at step {step_count}.")
+                    print(f"Lives: {previous_lives} -> {lives}")
+                    break
+
+                previous_lives = lives
 
         print("\nGym episode finished")
         print("Steps:", step_count)
